removed `/app` volume from dockerfile
Specifying a volume at `/app` (in the Dockerfile) makes it difficult to
use this image as a base image for derived images. Due to this, we have
removed the `VOLUME` instruction that mounts `/app` as a volume.

The `/app` path continues to function like it did before. As a resuly if
you want to mount your php application into the container, mount your
applucation source at `/app` using `-v
/path/on/application/source:/app`.

One advantage of mounting/installing your application at `/app` is that
the image will automatically update the ownership of the
files/directories in `/app` so that it is accessible by the `php-fpm`
process.